requirements.txt
PyQt6>=6.6.0,<7.0.0
PyMuPDF>=1.24.0,<2.0.0
Pillow>=10.0.0,<12.0.0
reportlab>=4.0.0,<5.0.0
pytesseract>=0.3.10,<1.0.0
